Final Commit
Ready for Review. Rename the files to their proper names. Added an event function which is called when the pages loads. Added the links to the html for the site to work properly.  Added in the event listeners to the buttons. Added in event listeners for the keys to work. Added a buttonClick function that is called when a button is clicked. Set the window value to what was called. Added in a calcKeys function that is called every time a key is pressed.